Replace CodeGenOutput trait with CompiledTrace.
